+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unrelased][v0.19.0](https://github.com/s3b4stian/linna-framework/compare/v0.18.0...v0.19.0) - 2017-XX-XX
 
+### Changed
+* `Linna\DI\Container` test updated
+* `Linna\DI\Resolver` merged with `Linna\DI\Container`
+
 ### Removed
 * `Linna\Cache\Exception\InvalidArgumentException` reminiscence of PSR simple-cache.